+++ core/src/plugins/kubernetes/results-cache-base.ts
@@ -23,8 +23,6 @@ import { tailString } from "../../util/string.js"
 import type { ContainerRunAction, ContainerTestAction } from "../container/config.js"
 import type { KubernetesRunAction, KubernetesTestAction } from "./kubernetes-type/config.js"
 import type { HelmPodRunAction, HelmPodTestAction } from "./helm/config.js"
-import { MAX_RUN_RESULT_LOG_LENGTH } from "./constants.js"
-import { getLocalActionResultsCacheDir, LocalResultCache } from "./results-cache-fs.js"
 
 export type CacheableAction = RunAction | TestAction
 
@@ -156,20 +154,4 @@ export abstract class AbstractResultCache<A extends CacheableAction, R extends C
   public abstract load(params: LoadResultParams<A>): Promise<R | undefined>
 
   public abstract store(params: StoreResultParams<A, R>): Promise<R | undefined>
-}
-
-let resultCache: LocalResultCache<CacheableRunAction | CacheableTestAction, CacheableResult> | undefined
-
-export function getResultCache(
-  gardenDirPath: string
-): LocalResultCache<CacheableRunAction | CacheableTestAction, CacheableResult> {
-  if (resultCache === undefined) {
-    resultCache = new LocalResultCache<CacheableRunAction | CacheableTestAction, CacheableResult>({
-      cacheDir: getLocalActionResultsCacheDir(gardenDirPath),
-      schemaVersion: currentResultSchemaVersion,
-      maxLogLength: MAX_RUN_RESULT_LOG_LENGTH,
-      resultValidator: kubernetesCacheableResultSchema.safeParse,
-    })
-  }
-  return resultCache
 }
