+++ core/src/plugins/kubernetes/test-results.ts
@@ -7,7 +7,8 @@
  */
 
 import type { TestActionHandler } from "../../plugin/action-types.js"
-import { getResultCache, toActionStatus } from "./results-cache-base.js"
+import { toActionStatus } from "./results-cache-base.js"
+import { getResultCache } from "./results-cache.js"
 
 // TODO: figure out how to get rid of the any cast
 export const k8sGetTestResult: TestActionHandler<"getResult", any> = async (params) => {
